refactor: move etf_asset_allocation from QuoteContext to FundamentalContext
Relocate the API and its types (AssetAllocationResponse / Group / Item,
HoldingDetail, ElementType) to the fundamental module across all
languages:

- Rust: types/method moved to fundamental, blocking wrapper moved to
  FundamentalContextSync
- Python: classes registered under fundamental module; method on
  FundamentalContext / AsyncFundamentalContext; openapi.pyi updated
- Node.js: etfAssetAllocation now on FundamentalContext;
  index.d.ts/index.js regenerated
- C: lb_quote_context_etf_asset_allocation →
  lb_fundamental_context_etf_asset_allocation; FFI types moved to
  fundamental_context; longbridge.h regenerated
- C++: method moved to fundamental::FundamentalContext; types moved to
  namespace fundamental
