+++ c/cbindgen.toml
@@ -304,13 +304,13 @@ cpp_compat = true
 "COptionVolumeStats" = "lb_option_volume_stats_t"
 "COptionVolumeDailyStat" = "lb_option_volume_daily_stat_t"
 "COptionVolumeDaily" = "lb_option_volume_daily_t"
+# FundamentalContext new types
 "CElementType" = "lb_element_type_t"
 "CLocaleName" = "lb_locale_name_t"
 "CHoldingDetail" = "lb_holding_detail_t"
 "CAssetAllocationItem" = "lb_asset_allocation_item_t"
 "CAssetAllocationGroup" = "lb_asset_allocation_group_t"
 "CAssetAllocationResponse" = "lb_asset_allocation_response_t"
-# FundamentalContext new types
 "CShareholderTopResponse" = "lb_shareholder_top_response_t"
 "CShareholderDetailResponse" = "lb_shareholder_detail_response_t"
 "CValuationHistoryPoint" = "lb_valuation_history_point_t"
@@ -447,10 +447,10 @@ include = [
     "CShortTradesItem", "CShortTradesResponse",
     "COptionVolumeStats",
     "COptionVolumeDailyStat", "COptionVolumeDaily",
+    # FundamentalContext new types
     "CElementType",
     "CLocaleName", "CHoldingDetail",
     "CAssetAllocationItem", "CAssetAllocationGroup", "CAssetAllocationResponse",
-    # FundamentalContext new types
     "CShareholderTopResponse", "CShareholderDetailResponse",
     "CValuationHistoryPoint", "CValuationComparisonItem", "CValuationComparisonResponse",
     # MarketContext new types
