+++ c/src/fundamental_context/enum_types.rs
@@ -80,3 +80,26 @@ pub enum CFinancialReportPeriod {
     #[c(remote = "ThreeQ")]
     FinancialReportPeriodThreeQ,
 }
+
+/// ETF asset allocation element type
+#[derive(Debug, Copy, Clone, Eq, PartialEq, CEnum)]
+#[c(remote = "longbridge::fundamental::types::ElementType")]
+#[allow(clippy::enum_variant_names)]
+#[repr(C)]
+pub enum CElementType {
+    /// Unknown
+    #[c(remote = "Unknown")]
+    ElementTypeUnknown,
+    /// Holdings
+    #[c(remote = "Holdings")]
+    ElementTypeHoldings,
+    /// Regional
+    #[c(remote = "Regional")]
+    ElementTypeRegional,
+    /// Asset class
+    #[c(remote = "AssetClass")]
+    ElementTypeAssetClass,
+    /// Industry
+    #[c(remote = "Industry")]
+    ElementTypeIndustry,
+}
