+++ c/src/quote_context/context.rs
@@ -1290,23 +1290,3 @@ pub unsafe extern "C" fn lb_quote_context_option_volume_daily(
         Ok(resp)
     });
 }
-
-/// Get ETF asset allocation (holdings / regional / asset class / industry).
-/// Returns `CAssetAllocationResponse`.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn lb_quote_context_etf_asset_allocation(
-    ctx: *const CQuoteContext,
-    symbol: *const c_char,
-    callback: CAsyncCallback,
-    userdata: *mut c_void,
-) {
-    use crate::{quote_context::types::CAssetAllocationResponseOwned, types::CCow};
-    let ctx_inner = (*ctx).ctx.clone();
-    let symbol = cstr_to_rust(symbol);
-    execute_async(callback, ctx, userdata, async move {
-        let resp: CCow<CAssetAllocationResponseOwned> = CCow::new(
-            CAssetAllocationResponseOwned::from(ctx_inner.etf_asset_allocation(symbol).await?),
-        );
-        Ok(resp)
-    });
-}
