+++ nodejs/src/quote/types.rs
@@ -1,5 +1,3 @@
-use std::collections::HashMap;
-
 use chrono::{DateTime, Utc};
 use longbridge::quote::SubFlags;
 use longbridge_nodejs_macros::{JsEnum, JsObject};
@@ -1662,125 +1660,6 @@ impl From<longbridge::quote::OptionVolumeDailyStat> for OptionVolumeDailyStat {
             total_put_open_interest: v.total_put_open_interest,
             total_call_open_interest: v.total_call_open_interest,
             put_call_open_interest_ratio: v.put_call_open_interest_ratio,
-        }
-    }
-}
-
-// ── etf_asset_allocation ──────────────────────────────────────────
-
-/// ETF asset allocation element type
-#[napi_derive::napi]
-#[derive(JsEnum, Debug, Hash, Eq, PartialEq, Copy, Clone)]
-#[js(remote = "longbridge::quote::ElementType")]
-pub enum ElementType {
-    /// Unknown
-    Unknown,
-    /// Holdings
-    Holdings,
-    /// Regional
-    Regional,
-    /// Asset class
-    AssetClass,
-    /// Industry
-    Industry,
-}
-
-/// Holding detail of an ETF asset allocation element (holdings only)
-#[napi_derive::napi(object)]
-#[derive(Debug, Clone)]
-pub struct HoldingDetail {
-    /// Industry ID
-    pub industry_id: String,
-    /// Industry name
-    pub industry_name: String,
-    /// Index counter ID (e.g. `BK/US/CP99000`)
-    pub index: String,
-    /// Index name
-    pub index_name: String,
-    /// Holding type (e.g. `E` for stock)
-    pub holding_type: String,
-    /// Holding type name
-    pub holding_type_name: String,
-}
-
-impl From<longbridge::quote::HoldingDetail> for HoldingDetail {
-    fn from(v: longbridge::quote::HoldingDetail) -> Self {
-        Self {
-            industry_id: v.industry_id,
-            industry_name: v.industry_name,
-            index: v.index,
-            index_name: v.index_name,
-            holding_type: v.holding_type,
-            holding_type_name: v.holding_type_name,
-        }
-    }
-}
-
-/// One element of an ETF asset allocation group
-#[napi_derive::napi(object)]
-#[derive(Debug, Clone)]
-pub struct AssetAllocationItem {
-    /// Element name
-    pub name: String,
-    /// Security code (holdings only, e.g. `NVDA`)
-    pub code: String,
-    /// Position ratio (e.g. `0.0861114`)
-    pub position_ratio: String,
-    /// Security symbol (holdings only, e.g. `NVDA.US`)
-    pub symbol: String,
-    /// Localized names (locale → name)
-    pub name_locales: HashMap<String, String>,
-    /// Holding detail (holdings only)
-    pub holding_detail: Option<HoldingDetail>,
-}
-
-impl From<longbridge::quote::AssetAllocationItem> for AssetAllocationItem {
-    fn from(v: longbridge::quote::AssetAllocationItem) -> Self {
-        Self {
-            name: v.name,
-            code: v.code,
-            position_ratio: v.position_ratio,
-            symbol: v.symbol,
-            name_locales: v.name_locales,
-            holding_detail: v.holding_detail.map(Into::into),
-        }
-    }
-}
-
-/// One ETF asset allocation group (grouped by element type)
-#[napi_derive::napi(object)]
-#[derive(Debug, Clone)]
-pub struct AssetAllocationGroup {
-    /// Report date (e.g. `20260601`)
-    pub report_date: String,
-    /// Element type of this group
-    pub asset_type: ElementType,
-    /// Elements
-    pub lists: Vec<AssetAllocationItem>,
-}
-
-impl From<longbridge::quote::AssetAllocationGroup> for AssetAllocationGroup {
-    fn from(v: longbridge::quote::AssetAllocationGroup) -> Self {
-        Self {
-            report_date: v.report_date,
-            asset_type: v.asset_type.into(),
-            lists: v.lists.into_iter().map(Into::into).collect(),
-        }
-    }
-}
-
-/// ETF asset allocation response
-#[napi_derive::napi(object)]
-#[derive(Debug, Clone)]
-pub struct AssetAllocationResponse {
-    /// Asset allocation groups
-    pub info: Vec<AssetAllocationGroup>,
-}
-
-impl From<longbridge::quote::AssetAllocationResponse> for AssetAllocationResponse {
-    fn from(v: longbridge::quote::AssetAllocationResponse) -> Self {
-        Self {
-            info: v.info.into_iter().map(Into::into).collect(),
         }
     }
 }
