+++ python/src/fundamental/context.rs
@@ -198,4 +198,14 @@ impl FundamentalContext {
             .map_err(ErrorNewType)?
             .into())
     }
+
+    /// Get ETF asset allocation (holdings / regional / asset class /
+    /// industry).
+    fn etf_asset_allocation(&self, symbol: String) -> PyResult<AssetAllocationResponse> {
+        Ok(self
+            .ctx
+            .etf_asset_allocation(symbol)
+            .map_err(ErrorNewType)?
+            .into())
+    }
 }
