+++ python/src/quote/context_async.rs
@@ -895,19 +895,4 @@ impl AsyncQuoteContext {
         })
         .map(|b| b.unbind())
     }
-
-    /// Get ETF asset allocation (holdings / regional / asset class /
-    /// industry). Returns awaitable.
-    fn etf_asset_allocation(&self, py: Python<'_>, symbol: String) -> PyResult<Py<PyAny>> {
-        let ctx = self.ctx.clone();
-        pyo3_async_runtimes::tokio::future_into_py(py, async move {
-            let r: crate::quote::types::AssetAllocationResponse = ctx
-                .etf_asset_allocation(symbol)
-                .await
-                .map_err(ErrorNewType)?
-                .into();
-            Ok(r)
-        })
-        .map(|b| b.unbind())
-    }
 }
